layout and renderer code refactor
